+++ src/main/java/pvpmode/command/PvPCommandConfig.java
@@ -36,7 +36,8 @@ public class PvPCommandConfig extends AbstractPvPCommand
     private void displayConfiguration (ICommandSender sender)
     {
         PvPUtils.postChatLines (sender, EnumChatFormatting.GREEN, "--- PvP Mode Configuration ---");
-        postConfigEntry (sender, "Warmup", Integer.toString (PvPMode.warmup) + "s");
+        postConfigEntry (sender, "Warmup off-on", Integer.toString (PvPMode.warmup) + "s");
+        postConfigEntry (sender, "Warmup on-off", Integer.toString (PvPMode.warmupOff) + "s");
         postConfigEntry (sender, "Cooldown", Integer.toString (PvPMode.cooldown) + "s");
         postConfigEntry (sender, "Radar enabled", Boolean.toString (PvPMode.radar));
         postConfigEntry (sender, "Distance round factor", Integer.toString (PvPMode.roundFactor));